+++ SistemaBliss.EN/Departamento.cs
@@ -8,7 +8,7 @@ namespace SistemaBliss.EN
 {
     public class Departamento
     {
-        public int IdDepartamento { get; set; }
+        public byte IdDepartamento { get; set; }
         public string Nombre { get; set; }
        
     }
