+++ SistemaBliss.EN/Estado.cs
@@ -8,7 +8,7 @@ namespace SistemaBliss.EN
 {
     public class Estado
     {
-        public int IdEstado { get; set; }
+        public byte IdEstado { get; set; }
         public string Nombre { get; set; }
     }
 }
